Improve error handling and search functionality
This commit improves the application's error handling by displaying the error messages in the DetailViewState.Error class to aid in troubleshooting.

Additionally, it enhances the search functionality in the FavoritesUI class to also include descriptions, thus broadening the search scope.

This commit also includes changes to handle large images by opening them in a URI handler and displaying an error message if the image is too large.

Lastly, the Kamel-Image library has been updated to version 0.8.2 for optimal performance.