Update Kotlin version and refactor border styling
This commit updates the Kotlin version from 1.9.20 to 1.9.23 and the Compose version from 1.5.10 to 1.6.1. The border styling of elements has also been changed from `AssistChipDefaults.assistChipBorder` to `BorderStroke`. A few variables have been commented out in the model classes and some code clean-up in Network and SettingsScreen are also included.