Add Scrollbars library and custom scrollbar
Added the GIGAMOLE:ComposeScrollbars library and included it in the build. Implemented a custom scrollbar component and added it to different target platforms including android and desktop. The new scrollbar is also incorporated into the UI for the application in the FavoritesUI section. This enhancement improves the user interface and navigation within the app.