Update CivitAiPagingSource and refactor FavoritesDatabase
Added early return clause in `getRefreshKey` method of CivitAiPagingSource.kt file to improve function workflow. Refactored FavoritesDatabase.kt by replacing `FavoriteList` class with `dateAdded` field on all subtypes of `FavoriteModel`. The refactoring simplifies the database schema and the `dateAdded` field will allow the tracking of when a favorite was added.