It's not really working but I'll wait until more better support
Update ViewModels and Screens to use SavedStateHandle

Switched the view models and screens to use a SavedStateHandle for saving and retrieving the state. The changes were made in 'CivitAiDetailViewModel', 'CivitAiUserViewModel', and 'CivitAiModelImagesViewModel'. This change enhances the robustness of the app in the face of configuration changes such as screen rotations. The required libraries for using `SavedStateHandle` in compose were also added to the build.gradle file.